moved the apscheduler background task function to separate file and folder in /jobs/background_tasks.py;
